+++ pymodbus/client/sync.py
@@ -1,5 +1,6 @@
 import socket
 import serial
+import time
 
 from pymodbus.constants import Defaults
 from pymodbus.factory import ClientDecoder
@@ -231,7 +232,7 @@ class ModbusUdpClient(BaseModbusClient):
             family = ModbusUdpClient._get_address_family(self.host)
             self.socket = socket.socket(family, socket.SOCK_DGRAM)
             self.socket.settimeout(self.timeout)
-        except socket.error, ex:
+        except socket.error as ex:
             _logger.error('Unable to create udp socket %s' % ex)
             self.close()
         return self.socket != None
@@ -305,6 +306,9 @@ class ModbusSerialClient(BaseModbusClient):
         self.parity   = kwargs.get('parity',   Defaults.Parity)
         self.baudrate = kwargs.get('baudrate', Defaults.Baudrate)
         self.timeout  = kwargs.get('timeout',  Defaults.Timeout)
+        if self.method == "rtu":
+            self._last_frame_end = 0.0
+            self._silent_interval = 3.5 * (1 + 8 + 2) / self.baudrate
 
     @staticmethod
     def __implementation(method):
@@ -330,9 +334,11 @@ class ModbusSerialClient(BaseModbusClient):
             self.socket = serial.Serial(port=self.port, timeout=self.timeout,
                 bytesize=self.bytesize, stopbits=self.stopbits,
                 baudrate=self.baudrate, parity=self.parity)
-        except serial.SerialException, msg:
+        except serial.SerialException as msg:
             _logger.error(msg)
             self.close()
+        if self.method == "rtu":
+            self._last_frame_end = time.time()
         return self.socket != None
 
     def close(self):
@@ -345,13 +351,40 @@ class ModbusSerialClient(BaseModbusClient):
     def _send(self, request):
         ''' Sends data on the underlying socket
 
+        If receive buffer still holds some data then flush it.
+
+        Sleep if last send finished less than 3.5 character
+        times ago.
+
         :param request: The encoded request to send
         :return: The number of bytes written
         '''
         if not self.socket:
             raise ConnectionException(self.__str__())
         if request:
-            return self.socket.write(request)
+            ts = time.time()
+            if self.method == "rtu":
+                if ts < self._last_frame_end + self._silent_interval:
+                    _logger.debug("will sleep to wait for 3.5 char")
+                    time.sleep(self._last_frame_end + self._silent_interval - ts)
+
+            try:
+                in_waiting = "in_waiting" if hasattr(self.socket, "in_waiting") else "inWaiting"
+                if in_waiting == "in_waiting":
+                    waitingbytes = getattr(self.socket, in_waiting)
+                else:
+                    waitingbytes = getattr(self.socket, in_waiting)()
+                if waitingbytes:
+                    result = self.socket.read(waitingbytes)
+                    if _logger.isEnabledFor(logging.WARNING):
+                        _logger.warning("cleanup recv buffer before send: " + " ".join([hex(ord(x)) for x in result]))
+            except NotImplementedError:
+                pass
+
+            size = self.socket.write(request)
+            if self.method == "rtu":
+                self._last_frame_end = time.time()
+            return size
         return 0
 
     def _recv(self, size):
@@ -362,7 +395,10 @@ class ModbusSerialClient(BaseModbusClient):
         '''
         if not self.socket:
             raise ConnectionException(self.__str__())
-        return self.socket.read(size)
+        result = self.socket.read(size)
+        if self.method == "rtu":
+            self._last_frame_end = time.time()
+        return result
 
     def __str__(self):
         ''' Builds a string representation of the connection