+++ pymodbus/internal/ptwisted.py
@@ -1,9 +1,11 @@
 '''
 A collection of twisted utility code
 '''
-from twisted.cred import portal, checkers
-from twisted.conch import manhole, manhole_ssh
-from twisted.conch.insults import insults
+from pymodbus.compat import IS_PYTHON2, IS_PYTHON3
+if IS_PYTHON2:
+    from twisted.cred import portal, checkers
+    from twisted.conch import manhole, manhole_ssh
+    from twisted.conch.insults import insults
 
 #---------------------------------------------------------------------------#
 # Logging
@@ -23,6 +25,8 @@ def InstallManagementConsole(namespace, users={'admin': 'admin'}, port=503):
     :param users: The users to login with
     :param port: The port to host the server on
     '''
+    if IS_PYTHON3:
+        raise NotImplemented("This code currently doesn't work on python3")
     from twisted.internet import reactor
 
     def build_protocol():