+++ pymodbus/mei_message.py
@@ -10,6 +10,7 @@ from pymodbus.pdu import ModbusResponse
 from pymodbus.device import ModbusControlBlock
 from pymodbus.device import DeviceInformationFactory
 from pymodbus.pdu import ModbusExceptions as merror
+from pymodbus.compat import iteritems, byte2int
 
 _MCB = ModbusControlBlock()
 
@@ -96,7 +97,7 @@ class ReadDeviceInformationResponse(ModbusResponse):
         :returns: The number of bytes in the response.
         '''
         size  = 8 # skip the header information
-        count = struct.unpack('>B', buffer[7])[0]
+        count = byte2int(buffer[7])
 
         while count > 0:
             _, object_length = struct.unpack('>BB', buffer[size:size+2])
@@ -129,10 +130,9 @@ class ReadDeviceInformationResponse(ModbusResponse):
             self.read_code, self.conformity, self.more_follows,
             self.next_object_id, self.number_of_objects)
 
-        for (object_id, data) in self.information.iteritems():
+        for (object_id, data) in iteritems(self.information):
             packet += struct.pack('>BB', object_id, len(data))
-            packet += data
-
+            packet += data.encode()
         return packet
 
     def decode(self, data):