+++ test/test_client_async.py
@@ -1,10 +1,13 @@
 #!/usr/bin/env python
 import unittest
-from mock import Mock
+from pymodbus.compat import IS_PYTHON3
+if IS_PYTHON3: # Python 3
+    from unittest.mock import patch, Mock
+else: # Python 2
+    from mock import patch, Mock
 from pymodbus.client.async import ModbusClientProtocol, ModbusUdpClientProtocol
 from pymodbus.client.async import ModbusClientFactory
 from pymodbus.exceptions import ConnectionException
-from pymodbus.exceptions import ParameterException
 from pymodbus.transaction import ModbusSocketFramer
 from pymodbus.bit_read_message import ReadCoilsRequest, ReadCoilsResponse
 
@@ -58,7 +61,7 @@ class AsynchronousClientTest(unittest.TestCase):
         protocol = ModbusClientProtocol()
         protocol.connectionMade()
         out = []
-        data = '\x00\x00\x12\x34\x00\x06\xff\x01\x01\x02\x00\x04'
+        data = b'\x00\x00\x12\x34\x00\x06\xff\x01\x01\x02\x00\x04'
 
         # setup existing request
         d = protocol._buildResponse(0x00)
@@ -130,7 +133,7 @@ class AsynchronousClientTest(unittest.TestCase):
         ''' Test the udp client protocol data received '''
         protocol = ModbusUdpClientProtocol()
         out = []
-        data = '\x00\x00\x12\x34\x00\x06\xff\x01\x01\x02\x00\x04'
+        data = b'\x00\x00\x12\x34\x00\x06\xff\x01\x01\x02\x00\x04'
         server = ('127.0.0.1', 12345)
 
         # setup existing request