+++ test/test_pdu.py
@@ -2,6 +2,7 @@
 import unittest
 from pymodbus.pdu import *
 from pymodbus.exceptions import *
+from pymodbus.compat import iteritems
 
 class SimplePduTest(unittest.TestCase):
     '''
@@ -39,7 +40,7 @@ class SimplePduTest(unittest.TestCase):
 
         result = self.exception.encode()
         self.exception.decode(result)
-        self.assertEqual(result, '\x01')
+        self.assertEqual(result, b'\x01')
         self.assertEqual(self.exception.exception_code, 1)
 
     def testRequestExceptionFactory(self):
@@ -47,31 +48,31 @@ class SimplePduTest(unittest.TestCase):
         request = ModbusRequest()
         request.function_code = 1
         errors = dict((ModbusExceptions.decode(c), c) for c in range(1,20))
-        for error, code in errors.iteritems():
+        for error, code in iteritems(errors):
             result = request.doException(code)
             self.assertEqual(str(result), "Exception Response(129, 1, %s)" % error)
 
     def testCalculateRtuFrameSize(self):
         ''' Test the calculation of Modbus/RTU frame sizes '''
         self.assertRaises(NotImplementedException,
-                          ModbusRequest.calculateRtuFrameSize, "")
+                          ModbusRequest.calculateRtuFrameSize, b'')
         ModbusRequest._rtu_frame_size = 5
-        self.assertEqual(ModbusRequest.calculateRtuFrameSize(""), 5)
+        self.assertEqual(ModbusRequest.calculateRtuFrameSize(b''), 5)
         del ModbusRequest._rtu_frame_size
 
         ModbusRequest._rtu_byte_count_pos = 2
         self.assertEqual(ModbusRequest.calculateRtuFrameSize(
-            "\x11\x01\x05\xcd\x6b\xb2\x0e\x1b\x45\xe6"), 0x05 + 5)
+            b'\x11\x01\x05\xcd\x6b\xb2\x0e\x1b\x45\xe6'), 0x05 + 5)
         del ModbusRequest._rtu_byte_count_pos
         
         self.assertRaises(NotImplementedException,
-                          ModbusResponse.calculateRtuFrameSize, "")
+                          ModbusResponse.calculateRtuFrameSize, b'')
         ModbusResponse._rtu_frame_size = 12
-        self.assertEqual(ModbusResponse.calculateRtuFrameSize(""), 12)
+        self.assertEqual(ModbusResponse.calculateRtuFrameSize(b''), 12)
         del ModbusResponse._rtu_frame_size
         ModbusResponse._rtu_byte_count_pos = 2
         self.assertEqual(ModbusResponse.calculateRtuFrameSize(
-            "\x11\x01\x05\xcd\x6b\xb2\x0e\x1b\x45\xe6"), 0x05 + 5)
+            b'\x11\x01\x05\xcd\x6b\xb2\x0e\x1b\x45\xe6'), 0x05 + 5)
         del ModbusResponse._rtu_byte_count_pos